tests: rename testappgwt2 to testapp.gwt2 and testappj2cl to testapp.j2cl
