+++ tests/tests-app-wrapper-gwt2/src/main/module.gwt.xml
@@ -1,5 +1,5 @@
 <!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.8.0//EN"
   "http://www.gwtproject.org/doctype/2.8.0/gwt-module.dtd">
 <module rename-to="VueGwtTestsAppGwt2">
-  <entry-point class="com.axellience.vuegwt.testsappgwt2.client.VueGwtTestsAppGwt2"/>
+  <entry-point class="com.axellience.vuegwt.testsapp.gwt2.client.VueGwtTestsAppGwt2"/>
 </module>
