+++ tests/tests-app-wrapper-j2cl/src/main/java/com/axellience/vuegwt/testsapp/j2cl/VueGwtTestsAppJ2CL.java
@@ -1,4 +1,4 @@
-package com.axellience.vuegwt.testsappj2cl;
+package com.axellience.vuegwt.testsapp.j2cl;
 
 import com.axellience.vuegwt.core.client.VueGWT;
 import com.axellience.vuegwt.testsapp.client.VueGwtTestApp;
@@ -7,6 +7,6 @@ public class VueGwtTestsAppJ2CL {
 
   public void onModuleLoad() {
     VueGWT.init();
-    VueGwtTestApp.bootstrap();
+    VueGwtTestsApp.bootstrap();
   }
 }
