+++ tests/tests-app-wrapper-j2cl/src/main/java/com/axellience/vuegwt/testsapp/j2cl/VueGwtTestsAppJ2CL.native.js
